Drop the hack to run ./manage.py with a specific Python version
The file shipped in python-django has a python shebang and the file
in python3-django has a python3 shebang.

We want to know it when it's no longer the case... so instead of
working around an issue that was present for some unknown reason
in 1.9-1, we let it fail (it's tested to work with a clean build
of 1.9-2).